requirements.txt
streamlit 
langchain-groq 
langchain-community 
langchain-ollama
faiss-cpu 
python-dotenv
fpdf2